requirements.txt
asgiref==3.8.1
attrs==25.3.0
coverage==7.9.1
Django==5.2.3
django-filter==25.1
django-stubs==4.2.7
django-stubs-ext==5.2.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mypy==1.8.0
mypy_extensions==1.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
ruff==0.1.6
sqlparse==0.5.3
types-pytz==2025.2.0.20250516
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.0
uritemplate==4.2.0
